% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls_authentification.R
\name{check_ns_api_keys}
\alias{check_ns_api_keys}
\title{Are the username and password set?}
\usage{
check_ns_api_keys()
}
\description{
A function to see if your username and password are set for nsapi.
Note that I have no clue if the password or username are correct,
that is up to you.
}
\details{
For more details see the help vignette:
\code{vignette("basic_use_nsapi_package", package = "nsapi")}
and get your keys here:
\url{https://www.ns.nl/ews-aanvraagformulier/?0}
}
