% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls_authentification.R
\name{datetime}
\alias{datetime}
\title{Translate date and time into datetime stamp to use in api call}
\usage{
datetime(date, time)
}
\arguments{
\item{date}{date in iso format (year-month-day): 2018-12-29 for example}

\item{time}{time in standard format (HH:MM): 20:21}
}
\description{
Translate date and time into datetime stamp to use in api call
}
