% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls_authentification.R
\name{disruptions_and_maintenance}
\alias{disruptions_and_maintenance}
\title{Get disruptions and engineering work}
\usage{
disruptions_and_maintenance(station = NULL, actual = TRUE,
  unplanned = TRUE)
}
\arguments{
\item{station}{station names need to be in Dutch and the NS webservic also accepts short versions:f.i. Groningen or GN}

\item{actual}{TRUE or FALSE indicator of the current disruptions must be returned This includes both unscheduled disruptions at the moment of the request, as well as engineering work scheduled to take place within two hours of the request.}

\item{unplanned}{TRUE or FALSE indicator of the scheduled engineering work for the next two weeks must be returned.}
}
\description{
\itemize{
\item current disruptions  (=unscheduled disruptions + current engineering work)
\item scheduled engineering work(=scheduled engineering work)
\item current disruptions for a specific station (=unscheduled disruptions + current engineering work)
}
}
