% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{General XML download function with cache option}
\usage{
download(url, ns.user = auth_cache$NS_USER, ns.pass = auth_cache$NS_PASS,
  use.cache = T, ...)
}
\description{
General XML download function with cache option
}

