% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls_authentification.R
\name{get_stationlist}
\alias{get_stationlist}
\title{Get a complete list of all the stations}
\usage{
get_stationlist()
}
\description{
This function should not be called too often.
The stations will probably not change a lot so it might
be better to save it as a dataframe in your local environment for further use.
}
\details{
The dataframe consists of all the Dutch stations, many German and Belgian stations and
bigger stations in other countries in Europe.
\url{https://www.ns.nl/en/travel-information/ns-api/documentation-station-list.html}
}
