% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls_authentification.R
\name{ns_api_check_keys}
\alias{ns_api_check_keys}
\title{Are the username and password set?}
\usage{
ns_api_check_keys()
}
\description{
A function to see if your username and password are set for nsapi.
Note that I have no clue if the password or username are correct,
that is up to you.
}
