% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsapi-package.R
\docType{package}
\name{nsapi-package}
\alias{nsapi}
\alias{nsapi-package}
\title{nsapi: Connect to the NS (Dutch Railways) api}
\description{
Something
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RMHogervorst/nsapi}
  \item Report bugs at \url{https://github.com/RMHogervorst/nsapi/issues}
}

}
\author{
\strong{Maintainer}: Roel M. Hogervorst \email{hogervorst.rm@gmail.com} (0000-0001-7509-0328)

}
\keyword{internal}
