% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls_authentification.R
\name{travel_advise}
\alias{travel_advise}
\title{Get travel advise from one station to another station}
\usage{
travel_advise(fromStation, toStation, dateTime = NULL, departure = TRUE,
  yearCard = FALSE, hslAllowed = FALSE, previousAdvices = 4,
  nextAdvices = 4)
}
\arguments{
\item{fromStation}{the station to start from, for instance "Rotterdam Centraal"}

\item{toStation}{the station to end, for instance "Utrecht Centraal"}

\item{dateTime}{defaults to current time, but you can use a different one: f.i. 2012-02-21T15:50, You can also use the \code{datetime()} function}

\item{departure}{is the datetime the start or end time? do you want to depart on that date or arrive, defaults to departure}

\item{yearCard}{if you have a NS year card (jaarabonnement) some trips will be different}

\item{hslAllowed}{use of the highspeed train}

\item{previousAdvices}{how many advices do you want before the time}

\item{nextAdvices}{how many advises do you want after}
}
\description{
This is equivalent to the \href{https://www.ns.nl/reisplanner#/}{NS reisplanner}, you give
in a from and to station time, departure or arrival, and optionally if you have
a NS year card (has effect on some travels). You can also specify how many trips
before and after the chosen time you want to call.
}
\details{
Some things to consider: station names need to be in Dutch and the NS webservice also
accepts shortened versions: "Utrecht Centraal" and "ut" is apparently the same.
Station names can be found with the \code{stationlist()} call.

Although the documentation \url{https://www.ns.nl/en/travel-information/ns-api/documentation-travel-recommendations.html} is in
english, the returns are all in Dutch. And I keep the results in English.
}
